feat: UserRepository 인터페이스 구현
- 의존성 주입, 다형성을 활용하기 위해 UserRepository 인터페이스 구현
