+++ src/main/java/codesquad/springcafe/domain/user/MemoryUserRepository.java
@@ -1,0 +1,36 @@
+package codesquad.springcafe.domain.user;
+
+import org.springframework.stereotype.Repository;
+
+import java.util.Map;
+import java.util.concurrent.ConcurrentHashMap;
+
+@Repository
+public class UserMemoryRepository implements UserRepository {
+
+    private final Map<String, User> users;
+
+    public UserMemoryRepository() {
+        this.users = new ConcurrentHashMap<>();
+    }
+
+    @Override
+    public void save(User user) {
+        user.setId((long) (users.size() + 1));
+        users.put(user.getUserId(), user);
+    }
+
+    @Override
+    public void update(User user) {
+        users.put(user.getUserId(), user);
+    }
+
+    @Override
+    public User findByUserId(String userId) {
+        return users.get(userId);
+    }
+
+    public Map<String, User> getUsers() {
+        return users;
+    }
+}
